+++ src/Commands/PublishCmd.ts
@@ -23,24 +23,33 @@ export class PublishCmd {
     constructor() {
     }
 
-    public ExecutePublishCmd(): void {
+    public ExecutePublishCmd(args): void {
 
-        if (ValidationUtility.WorkspaceValidation()) {
-            let rootFolders = ValidationUtility.SelectRootPath();
-            if (rootFolders.size > 1) {
-                // Select the workspace folder.
-                QuickPickUtility.ShowQuickPick(Array.from(rootFolders.keys()), StringUtility.SelectWorkspaceFolder)
-                    .then(response => {
-                        if (typeof response != StringUtility.Undefined) {
-                            let rootPath = rootFolders.get(response);
-                            PublishCmd.GetProjectPath(rootPath);
-                        }
-                    });
+        if (typeof args == StringUtility.Undefined) {
+            if (ValidationUtility.WorkspaceValidation()) {
+                let rootFolders = ValidationUtility.SelectRootPath();
+                if (rootFolders.size > 1) {
+                    // Select the workspace folder.
+                    QuickPickUtility.ShowQuickPick(Array.from(rootFolders.keys()), StringUtility.SelectWorkspaceFolder)
+                        .then(response => {
+                            if (typeof response != StringUtility.Undefined) {
+                                let rootPath = rootFolders.get(response);
+                                PublishCmd.GetProjectPath(rootPath);
+                            }
+                        });
+                }
+                else {
+                    let rootPath = rootFolders.values().next().value;
+                    PublishCmd.GetProjectPath(rootPath);
+                }
             }
-            else {
-                let rootPath = rootFolders.values().next().value;
-                PublishCmd.GetProjectPath(rootPath);
-            }
+        }
+        else {
+
+            var csProjName = args.fsPath.substring(args.fsPath.lastIndexOf('\\') + 1);
+            var folderPath = args.fsPath.toString().replace(csProjName, "")
+
+            PublishCmd.PublishProject(folderPath, csProjName);
         }
     }