+++ CHANGELOG.md
@@ -1,14 +1,15 @@
 # Dotnet Core Essentials Change Log
 
-## 0.0.1 - 2017-12-10
+## 0.0.1 - 2017-12-25
 
 ### Added
 
-- Creating Console, Web-API, MVC & ClassLibrary projects in core 2.0 framework
-- Create Angular, React, React-Redux templates
+- Creating Console, Web-API, MVC, Angular, React, React-Redux & ClassLibrary projects in core 2.0 framework and standard framework.
 - Adding project/dll reference
 - Publishing project
 - Selecting a startup project
+- Build/Clean project and solution
+- Adding nuget packages
 - keyboard shortcuts
   - `Ctrl+Alt+N` - New project
   - `Ctrl+Alt+A` - Add project/dll reference